Chris_solution/DIC_Ex3_Valentin/My_project/requirements.txt
Flask
pybase64
jsonpickle
Pillow
numpy
opencv-python
fastapi
tflite
https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.9.0-py3-none-any.whl
tflite-runtime
tensorflow_hub
